Move value specific algorithms
