+++ CheckDigits.Net.Tests.Benchmarks/AbaRtnAlgorithmValidateBenchmarks.cs
@@ -1,5 +1,7 @@
 ﻿// Ignore Spelling: Aba Rtn
 
+using CheckDigits.Net.ValueSpecificAlgorithms;
+
 namespace CheckDigits.Net.Tests.Benchmarks;
 
 [MemoryDiagnoser]
