+++ CheckDigits.Net.Tests.Benchmarks/IsinAlgorithmValidateBenchmarks.cs
@@ -1,5 +1,7 @@
 ﻿// Ignore Spelling: Isin
 
+using CheckDigits.Net.ValueSpecificAlgorithms;
+
 namespace CheckDigits.Net.Tests.Benchmarks;
 
 [MemoryDiagnoser]
