+++ CheckDigits.Net.Tests.Benchmarks/NpiAlgorithmValidateBenchmarks.cs
@@ -1,5 +1,7 @@
 ﻿// Ignore Spelling: Npi
 
+using CheckDigits.Net.ValueSpecificAlgorithms;
+
 namespace CheckDigits.Net.Tests.Benchmarks;
 
 [MemoryDiagnoser]
