+++ CheckDigits.Net.Tests.Unit/GlobalUsings.cs
@@ -4,4 +4,5 @@ global using Xunit;
 
 global using CheckDigits.Net.Iso7064;
 global using CheckDigits.Net.Utility;
+global using CheckDigits.Net.ValueSpecificAlgorithms;
 global using CheckDigits.Net.Tests.Unit.TestData;
