+++ CheckDigits.Net/GlobalUsings.cs
@@ -1,1 +1,3 @@
-﻿global using CheckDigits.Net.Utility;
+﻿global using CheckDigits.Net.Iso7064;
+global using CheckDigits.Net.Utility;
+global using CheckDigits.Net.ValueSpecificAlgorithms;
